brandCRUD

+++ src/main/java/com/perz/carrentalapp/service/RoleService.java
@@ -2,26 +2,16 @@ package com.perz.carrentalapp.service;
 
 import com.perz.carrentalapp.model.Role;
 import com.perz.carrentalapp.repositories.RoleRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.perz.carrentalapp.util.exceptions.RoleNotFoundException;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
-
-
 @Service
+@AllArgsConstructor
 @Transactional(readOnly = true)
 public class RoleService {
     private final RoleRepository roleRepository;
-
-    @Autowired
-    public RoleService(RoleRepository roleRepository) {
-        this.roleRepository = roleRepository;
-    }
-
-    public Optional<Role> findByName(String name) {
-        return roleRepository.findByName(name);
-    }
 
     @Transactional
     public void create(Role role) {
@@ -29,13 +19,24 @@ public class RoleService {
     }
 
     public Role findOne(Long id) {
-        return roleRepository.findById(id).orElse(null);
+
+        Role role = roleRepository.findById(id).orElse(null);
+
+        if (role == null) {
+            throw new RoleNotFoundException("Role not found with id: " + id);
+        }
+
+        return role;
     }
 
     @Transactional
     public void update(Long id, Role role) {
 
-        Role roleToBeUpdate = roleRepository.findById(id).get();
+        Role roleToBeUpdate = roleRepository.findById(id).orElse(null);
+
+        if (roleToBeUpdate == null) {
+            throw new RoleNotFoundException("Role not found with id: " + id);
+        }
 
         roleToBeUpdate.setName(role.getName());
 
@@ -45,7 +46,11 @@ public class RoleService {
     @Transactional
     public void delete(Long id) {
 
-        Role roleToBeDelete = roleRepository.findById(id).get();
+        Role roleToBeDelete = roleRepository.findById(id).orElse(null);
+
+        if (roleToBeDelete == null) {
+            throw new RoleNotFoundException("Role not found with id: " + id);
+        }
 
         roleRepository.delete(roleToBeDelete);
     }
